requirements.txt
torch>=2.1
transformers>=4.40
peft>=0.10
bitsandbytes>=0.43
accelerate>=0.30
datasets
trl
rouge-score
numpy
pandas
pyyaml
tqdm
